+++ brick/__brick__/packages/{{project_name.snakeCase()}}/lib/theme/subthemes/icon_button_theme.dart
@@ -1,0 +1,33 @@
+import 'package:{{project_name.snakeCase()}}/tokens/colors.dart';
+import 'package:{{project_name.snakeCase()}}/tokens/sizes.dart';
+import 'package:flutter/material.dart';
+
+class FlexIconButtonTheme {
+  FlexIconButtonTheme._();
+
+  static IconButtonThemeData lightIconButtonTheme = IconButtonThemeData(
+    style: IconButton.styleFrom(
+      backgroundColor: Colors.transparent,
+      foregroundColor: FlexAppColorScheme.lightScheme.primary,
+      disabledBackgroundColor: Colors.transparent,
+      disabledForegroundColor: FlexAppColorScheme.lightScheme.disabled,
+      padding: const EdgeInsets.all(FlexSizes.sm),
+      shape: RoundedRectangleBorder(
+        borderRadius: BorderRadius.circular(FlexSizes.borderRadiusSm),
+      ),
+    ),
+  );
+
+  static IconButtonThemeData darkIconButtonTheme = IconButtonThemeData(
+    style: IconButton.styleFrom(
+      backgroundColor: Colors.transparent,
+      foregroundColor: FlexAppColorScheme.darkScheme.primary,
+      disabledBackgroundColor: Colors.transparent,
+      disabledForegroundColor: FlexAppColorScheme.darkScheme.disabled,
+      padding: const EdgeInsets.all(FlexSizes.sm),
+      shape: RoundedRectangleBorder(
+        borderRadius: BorderRadius.circular(FlexSizes.borderRadiusSm),
+      ),
+    ),
+  );
+}