+++ brick/__brick__/packages/{{project_name.snakeCase()}}/lib/theme/subthemes/outlined_button_theme.dart
@@ -1,3 +1,4 @@
+import 'package:{{project_name.snakeCase()}}/tokens/colors.dart';
 import 'package:flutter/material.dart';
 
 class FlexOutlinedButtonTheme {
@@ -7,10 +8,13 @@ class FlexOutlinedButtonTheme {
       OutlinedButtonThemeData(
     style: OutlinedButton.styleFrom(
       elevation: 0,
-      foregroundColor: Colors.black,
-      side: const BorderSide(color: Colors.blue),
-      textStyle: const TextStyle(
-          fontSize: 16, color: Colors.black, fontWeight: FontWeight.w600),
+      foregroundColor: FlexAppColorScheme.lightScheme.primary,
+      side: BorderSide(color: FlexAppColorScheme.lightScheme.info),
+      textStyle: TextStyle(
+        fontSize: 16,
+        color: FlexAppColorScheme.lightScheme.primary,
+        fontWeight: FontWeight.w600,
+      ),
       padding: const EdgeInsets.symmetric(vertical: 16, horizontal: 20),
       shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(14)),
     ),
@@ -19,10 +23,13 @@ class FlexOutlinedButtonTheme {
   static OutlinedButtonThemeData darkOutlinedButtonTheme =
       OutlinedButtonThemeData(
     style: OutlinedButton.styleFrom(
-      foregroundColor: Colors.white,
-      side: const BorderSide(color: Colors.blueAccent),
-      textStyle: const TextStyle(
-          fontSize: 16, color: Colors.white, fontWeight: FontWeight.w600),
+      foregroundColor: FlexAppColorScheme.darkScheme.primary,
+      side: BorderSide(color: FlexAppColorScheme.darkScheme.info),
+      textStyle: TextStyle(
+        fontSize: 16,
+        color: FlexAppColorScheme.darkScheme.primary,
+        fontWeight: FontWeight.w600,
+      ),
       padding: const EdgeInsets.symmetric(vertical: 16, horizontal: 20),
       shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(14)),
     ),