+++ brick/__brick__/packages/{{project_name.snakeCase()}}/lib/{{project_name.snakeCase()}}.dart
@@ -1,10 +1,22 @@
 // Widgets
-export 'widgets/add_to_cart_button.dart';
+export 'widgets/buttons/button.dart';
 export 'widgets/banner/banner.dart';
+export 'widgets/carousel/carousel.dart';
 export 'widgets/image/image.dart';
+export 'widgets/quantity_selector/quantity_selector.dart';
+export 'widgets/cards/productCard/shared/price.dart';
+export 'widgets/cards/productCard/shared/price_discount.dart';
+export 'widgets/gallery/gallery.dart';
+export 'widgets/image/tappable_image.dart';
+export 'widgets/cards/productCard/shared/badge.dart';
+export 'widgets/search/search.dart';
+
 // Themes
 export 'theme/flex_app_themes.dart';
 
 // Tokens
 export 'tokens/colors.dart';
 export 'tokens/sizes.dart';
+
+// Extensions
+export 'utils/extensions.dart';